Adiciona e atualiza arquivos RFs
