+++ order-service/src/main/java/com/scd/order/OrderRequest.java
@@ -1,0 +1,15 @@
+package com.scd.order;
+
+import java.util.List;
+
+public class OrderRequest {
+    private List<OrderItem> items;
+
+    public List<OrderItem> getItems() {
+        return items;
+    }
+
+    public void setItems(List<OrderItem> items) {
+        this.items = items;
+    }
+}
